refactor(shortcuts): move the session hook into @agenta/ui so /m can reuse it
The plan is for /m to become the whole app, not just the phone surface. Four of
the five shortcut pieces already sat in packages /m depends on. The hook did
not: it lived in web/oss, even though it imports only React and
@agenta/shared/utils and takes every action as a callback. Nothing about it was
desktop-specific.

It moves to @agenta/ui/shortcuts beside the keycaps, the sheet and the help
button, so a mobile host wires the same callbacks rather than reimplementing
anything. A phone never sends an Alt chord, so mounting it on a touch surface is
inert.

The old test covered two things at once, the hook and an app-layer rename
consumer that needs OSS atoms. It splits: the hook half moves to the package
(20 tests), the rename half stays in web/oss as useInlineRenameRequest.test.ts.

Storybook gains an On mobile story showing the approval card at phone width and
beside its desktop twin, so a keycap leaking onto touch is caught by looking.

Also fixes a story that threw at render. shortcutAria was used without its
import, which neither tsc nor eslint catches here: SWC transpiles stories
without type checking, so it built clean and failed only in the browser. My
regression sweep missed it too, because a thrown story renders Storybook's error
boundary as ordinary text. The sweep now reads that text. 60 renders, ten pages
across two themes and three widths, all clean.
