+++ web/packages/agenta-ui/src/shortcuts/useSessionShortcuts.ts
@@ -44,7 +44,12 @@ const steppedSession = <T extends {id: string}>(
 }
 
 /**
- * Session shortcuts for the agent playground: `Alt+1…9` jumps to the Nth open session, `Alt+Z` and
+ * Session shortcuts for a chat surface that keeps sessions in a strip. Every action arrives as a
+ * callback, so the hook knows nothing about its host: the desktop playground drives it today, and
+ * `/m` can drive the same one when it grows a keyboard surface. A phone never sends an Alt chord,
+ * so mounting it there is inert rather than harmful.
+ *
+ * The bindings: `Alt+1…9` jumps to the Nth open session, `Alt+Z` and
  * `Alt+X` step to the previous/next one (wrapping), `Alt+N` opens a new session, `Alt+W` closes the
  * active one, `Alt+R` renames it, `Alt+A` archives it, `Alt+K` searches, `Alt+C` toggles the config
  * panel, `Alt+O` toggles the files pane. Stop and approve live with the conversation that owns the
